backend/requirements.txt
# Core FastAPI and web framework dependencies
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.2

# Database and ORM
sqlalchemy==2.0.23

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
cryptography>=41.0.0
appwrite>=6.0.0

# Google OAuth
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib>=1.1.0

# File handling and validation
python-multipart>=0.0.6
pydantic[email]>=2.5.0
aiofiles==24.1.0
pillow==11.3.0

# Data science and ML for recommendations
pandas==2.3.2
numpy==2.3.2
scikit-learn==1.7.1
scipy==1.16.1
joblib==1.5.1
threadpoolctl==3.6.0

# HTTP and networking
httplib2==0.22.0
requests>=2.31.0
urllib3>=1.26.0
certifi>=2023.0.0
aiohttp==3.12.15
faker==33.1.0

# Date and time handling
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Utilities
click>=8.0.0
typing-extensions>=4.8.0
annotated-types>=0.6.0
pydantic-core>=2.14.0

# Email validation
email-validator>=2.0.0
dnspython>=2.0.0

# Cryptographic dependencies
cffi>=1.14.0
bcrypt>=4.0.0
rsa>=4.9.0
pyasn1>=0.6.0
pyasn1-modules>=0.4.0
ecdsa>=0.18.0

# Web server dependencies
h11>=0.16.0
anyio>=4.0.0
sniffio>=1.3.0
idna>=3.4

# JSON Web Token support
jose>=1.0.0

# Additional utilities
six>=1.16.0
greenlet>=3.0.0
cachetools>=5.3.0
